+++ pkg/cmd/start/bootstrap.go
@@ -7,6 +7,7 @@ import (
 	"os"
 	"runtime"
 	"strings"
+	"time"
 
 	osimagev1 "github.com/openshift/api/image/v1"
 	log "github.com/sirupsen/logrus"
@@ -34,6 +35,7 @@ import (
 	jaegertracingv1 "github.com/jaegertracing/jaeger-operator/apis/v1"
 	v1 "github.com/jaegertracing/jaeger-operator/apis/v1"
 	appsv1controllers "github.com/jaegertracing/jaeger-operator/controllers/appsv1"
+	esv1controllers "github.com/jaegertracing/jaeger-operator/controllers/elasticsearch"
 	jaegertracingcontrollers "github.com/jaegertracing/jaeger-operator/controllers/jaegertracing"
 	"github.com/jaegertracing/jaeger-operator/pkg/autodetect"
 	kafkav1beta2 "github.com/jaegertracing/jaeger-operator/pkg/kafka/v1beta2"
@@ -114,6 +116,7 @@ func bootstrap(ctx context.Context) manager.Manager {
 	detectNamespacePermissions(ctx, mgr)
 	performUpgrades(ctx, mgr)
 	setupControllers(ctx, mgr)
+	setupWebhooks(ctx, mgr)
 	detectOAuthProxyImageStream(ctx, mgr)
 	err = opmetrics.Bootstrap(ctx, namespace, mgr.GetClient())
 	if err != nil {
@@ -281,6 +284,10 @@ func createManager(ctx context.Context, cfg *rest.Config) manager.Manager {
 
 	namespace := viper.GetString(v1.ConfigWatchNamespace)
 
+	// see https://github.com/openshift/library-go/blob/4362aa519714a4b62b00ab8318197ba2bba51cb7/pkg/config/leaderelection/leaderelection.go#L104
+	leaseDuration := time.Second * 137
+	renewDeadline := time.Second * 107
+	retryPeriod := time.Second * 26
 	options := ctrl.Options{
 		Scheme:                 scheme,
 		MetricsBindAddress:     metricsAddr,
@@ -288,6 +295,9 @@ func createManager(ctx context.Context, cfg *rest.Config) manager.Manager {
 		HealthProbeBindAddress: probeAddr,
 		LeaderElection:         enableLeaderElection,
 		LeaderElectionID:       "31e04290.jaegertracing.io",
+		LeaseDuration:          &leaseDuration,
+		RenewDeadline:          &renewDeadline,
+		RetryPeriod:            &retryPeriod,
 		Namespace:              namespace,
 	}
 
@@ -352,6 +362,18 @@ func setupControllers(ctx context.Context, mgr manager.Manager) {
 
 	if err := appsv1controllers.NewDeploymentReconciler(client, clientReader, schema).SetupWithManager(mgr); err != nil {
 		setupLog.Error(err, "unable to create controller", "controller", "Deployment")
+		os.Exit(1)
+	}
+
+	if err := esv1controllers.NewReconciler(client, clientReader).SetupWithManager(mgr); err != nil {
+		setupLog.Error(err, "unable to create controller", "controller", "Elasticsearch")
+		os.Exit(1)
+	}
+}
+
+func setupWebhooks(_ context.Context, mgr manager.Manager) {
+	if err := (&v1.Jaeger{}).SetupWebhookWithManager(mgr); err != nil {
+		setupLog.Error(err, "unable to create webhook", "webhook", "Jaeger")
 		os.Exit(1)
 	}
 }