+++ pkg/metrics/instances.go
@@ -18,6 +18,7 @@ const metricPrefix = "jaeger_operator_instances"
 const agentStrategiesMetric = "agent_strategies"
 const storageMetric = "storage_types"
 const strategiesMetric = "strategies"
+const autoprovisioningMetric = "autoprovisioning"
 
 // This structure contains the labels associated with the instances and a counter of the number of instances
 type instancesView struct {
@@ -35,7 +36,10 @@ func (i *instancesView) reset() {
 }
 
 func (i *instancesView) Record(jaeger v1.Jaeger) {
-	i.Count[i.KeyFn(jaeger)]++
+	label := i.KeyFn(jaeger)
+	if label != "" {
+		i.Count[label]++
+	}
 }
 
 func (i *instancesView) Report(result metric.BatchObserverResult) {
@@ -115,6 +119,21 @@ func (i *instancesMetric) Setup(ctx context.Context) error {
 		return err
 	}
 	i.observations = append(i.observations, obs)
+
+	obs, err = newObservation(batch, autoprovisioningMetric,
+		"Number of instances using autoprovisioning",
+		"type",
+		func(jaeger v1.Jaeger) string {
+			if v1.ShouldInjectOpenShiftElasticsearchConfiguration(jaeger.Spec.Storage) {
+				return "elasticsearch"
+			}
+			return ""
+		})
+	if err != nil {
+		return err
+	}
+	i.observations = append(i.observations, obs)
+
 	return nil
 }