+++ pkg/strategy/controller.go
@@ -15,11 +15,6 @@ import (
 
 	v1 "github.com/jaegertracing/jaeger-operator/apis/v1"
 	"github.com/jaegertracing/jaeger-operator/pkg/cronjob"
-	"github.com/jaegertracing/jaeger-operator/pkg/storage"
-)
-
-const (
-	esCertGenerationScript = "./scripts/cert_generation.sh"
 )
 
 var (
@@ -129,6 +124,10 @@ func normalize(ctx context.Context, jaeger *v1.Jaeger) {
 	normalizeElasticsearch(&jaeger.Spec.Storage.Elasticsearch)
 	normalizeRollover(&jaeger.Spec.Storage.EsRollover)
 	normalizeUI(&jaeger.Spec)
+
+	if jaeger.Spec.Storage.Elasticsearch.Name == "" {
+		jaeger.Spec.Storage.Elasticsearch.Name = "elasticsearch"
+	}
 }
 
 func distributedStorage(storage v1.JaegerStorageType) bool {
@@ -146,7 +145,7 @@ func normalizeSparkDependencies(spec *v1.JaegerStorageSpec) {
 	// auto enable only for supported storages
 	if cronjob.SupportedStorage(spec.Type) &&
 		spec.Dependencies.Enabled == nil &&
-		!storage.ShouldDeployElasticsearch(*spec) &&
+		!v1.ShouldInjectOpenShiftElasticsearchConfiguration(*spec) &&
 		tlsIsNotEnabled {
 		trueVar := true
 		spec.Dependencies.Enabled = &trueVar